fix navbar login btn using context
