+++ components/Navbar.jsx
@@ -4,22 +4,27 @@ import Link  from "next/link";
 import axios from "axios";
 import { useEffect, useState} from 'react';
 import {useRouter} from "next/navigation";
+import { UserContext } from './userContext';
+import { useContext } from 'react';
 import Cookies from 'js-cookie';
 const Navbar = ({loggedIn}) => {
+  const {user,setUser}=useContext(UserContext);
   const [isLoggedIn, setIsLoggedIn] = useState(false)
-  useEffect(() => {
-    const getIsLoggedInFromCookies = () => {
-      const isLoggedInCookie = Cookies.get('isLoggedIn');
-      console.log(isLoggedInCookie);
-        setIsLoggedIn(isLoggedInCookie === 'true');
-    };
+//   useEffect(() => {
+//     const getIsLoggedInFromCookies = () => {
+//       const isLoggedInCookie =user;
+//       console.log(isLoggedInCookie);
+//         setIsLoggedIn(isLoggedInCookie);
+//     };
 
-    getIsLoggedInFromCookies();
-}, []);
+//     getIsLoggedInFromCookies();
+// }, []);
+
   const router = useRouter();
   const logout = async () => {
     try {
       await axios.get("/api/users/logout");
+      setUser(false);
       router.push("/login");
     } catch (error) {
       console.log(error);
@@ -35,7 +40,7 @@ const Navbar = ({loggedIn}) => {
         <Link href="/component"><div className="m-2 p-2">My Components</div></Link>
         <Link href="/about_us"><div className="m-2 p-2">About Us</div></Link>
           <button onClick={logout} className="w-fit text-white bg-primary-600 hover:bg-primary-700 focus:outline-none focus:ring-primary-300 font-medium rounded-lg text-sm py-2.5 px-4 text-center dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800">
-            {isLoggedIn ? "Logout" : "Login"}
+            {user? "Logout" : "Login"}
           </button>
         </div>
       </div>
