+++ src/app/layout.js
@@ -4,6 +4,7 @@ import { Navbar, Footer } from "@/components";
 import "./globals.css";
 import axios from "axios";
 import { useEffect } from "react";
+import { UserContextProvider } from "@/components/userContext";
 import { Toaster } from "react-hot-toast";
 // const inter = Inter({ subsets: ["latin"] });
 
@@ -28,9 +29,12 @@ export default function RootLayout({ children }) {
             },
           }}
         />
-        <Navbar />
-        {children}
-        <Footer />
+        <UserContextProvider>
+          {" "}
+          <Navbar />
+          {children}
+          <Footer />
+        </UserContextProvider>
       </body>
     </html>
   );
